Also load plugin dependencies in Ohai
This fixes #43

Loading all_plugins as we used to do, fails on a clean 10.9 Mavericks with 'Unable to find any JVMs matching version "(null)". No Java runtime present, try --request to install'. And running java with --requests sends you to a page on the Oracle site. Not what we want for Kitchenplan since we install Java using Chef later on.

The reason it fails on this, is because an Ohai plugin tests for Java. We only need the platform_family, so I only loaded the "platform" plugin. Now apparently I failed to detect that it broke and only installed the not platform dependant recipes since the Kitchenplan platform was empty. 

This is because in the platform plugin Ohai loads "#{os}::platform" but "os" was empty. This is because the "os" variable comes from the "os" plugin, but even though Ohai has a dependency system for plugins, the platform plugin does not require the os plugin. This is fixed in Ohai master, but not yet in the stable release.

- Master: https://github.com/opscode/ohai/blob/master/lib/ohai/plugins/platform.rb
- Stable: https://github.com/opscode/ohai/blob/6-stable/lib/ohai/plugins/platform.rb