Better handling of config directory conflicts between repositories.
In the case where a Kitchenplan application repo contains a config
directory _and_ a KITCHENPLAN_CONFIG_REPO environment variable is
passed into the go script, it will prefer the external repository.
If you want it to prefer the application's config directory, unset
the KITCHENPLAN_CONFIG_REPO environment variable.

I had to make these tweaks in order to test the go script in a pull
request in Travis.

Conflicts:
	go